+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 <!-- Adjusted classes functions scopes to `private` where possible -->
 <!-- Fixed an issue making it unable to initialize plugins correctly -->
+<!-- Removed email notification for a deleted worktime entry -->
 
 ## v7.2.1